Split up the code into more managable chunks (WIP)
