+++ piplicenses/cli/CompatibleArgumentParser.py
@@ -1,0 +1,66 @@
+# pip-licenses-cli
+#
+# MIT License
+#
+# Copyright (c) 2018 raimon
+# Copyright (c) 2025 stefan6419846
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from __future__ import annotations
+
+import argparse
+
+from .CustomNamespace import CustomNamespace
+
+import codecs
+from enum import Enum, auto
+
+from .OrderArg import OrderArg
+from .FormatArg import FormatArg
+from .CustomHelpFormatter import CustomHelpFormatter
+
+from typing import TYPE_CHECKING, Protocol, Type, cast
+
+if TYPE_CHECKING:  # pragma: no cover
+    from typing import Iterator, Optional, Sequence
+
+
+class CompatibleArgumentParser(argparse.ArgumentParser):
+    def parse_args(  # type: ignore[override]
+        self,
+        args: None | Sequence[str] = None,
+        namespace: None | CustomNamespace = None,
+    ) -> CustomNamespace:
+        args_ = cast(CustomNamespace, super().parse_args(args, namespace))
+        self._verify_args(args_)
+        return args_
+
+    def _verify_args(self, args: CustomNamespace) -> None:
+        if args.with_license_file is False and (args.no_license_path is True or args.with_notice_file is True):
+            self.error("'--no-license-path' and '--with-notice-file' require " "the '--with-license-file' option to be set")
+        if args.filter_strings is False and args.filter_code_page != "latin1":
+            self.error("'--filter-code-page' requires the '--filter-strings' " "option to be set")
+        try:
+            codecs.lookup(args.filter_code_page)
+        except LookupError:
+            self.error(
+                f"invalid code page {args.filter_code_page!r} given for '--filter-code-page, check "
+                "https://docs.python.org/3/library/codecs.html#standard-encodings for valid code pages"
+            )
